Re-import httpuv.
'httpuv' is a runtime dep of 'httr' for the OAuth dance. The other previously included runtime dep 'jsonlite' is now directly imported by httr. So no need to add explicitly as our package dep.

The reason why Travis did not catch this was because our Travis tests do not perform the OAuth dance.